requirements.txt
tgcrypto
asyncio
ffmpeg-python
py-tgcalls
pyrogram
youtube-search-python
yt-dlp
speedtest-cli
git+https://github.com/ytdl-org/youtube-dl@master
youtube-search
python-dotenv
dnspython
gitpython
aiofiles
aiohttp
requests
motor
psutil
future
wget
lyricsgenius
tgcrypto
pyrogram==1.4.16
py-tgcalls==0.9.0b1

wget
motor
pillow
psutil
asyncio
aiohttp
aiofiles
requests
lyricsgenius
speedtest-cli

dnspython
gitpython
ffmpeg-python
python-dotenv

yt-dlp
youtube-search
youtube-search-python
